arch-atmega/uart.c driver with fixed register values for baudrates
The way uart.c used to compute the register values for baudrates was buggy
for some values and did not use U2X1 bit in UCSR1A, which leads to a more
precise baudrate.
More simply, here I introduce an array with pre-defined settings for
the available baudrates. Values are taken from table 18-12 of atmega32u4
datasheet.
The unuseful baudrate 125000 is replaced by more common 115200.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>